Created a more organized file system and started an initial header menu. Added ui library
