+++ src/app/App.css
@@ -1,0 +1,8 @@
+.App {
+  text-align: center;
+}
+
+.header {
+  background-color: #222;
+  padding-bottom: 50px;
+}
